requirements.txt
torch~=1.4.0
torchvision~=0.5.0
numpy~=1.18.1
Pillow~=6.2.1
matplotlib~=3.1.3
torchtext~=0.6.0